refactor: Rename "events" package to "event"
