+++ common/src/main/java/com/wynntils/core/chat/tabs/ChatTab.java
@@ -5,7 +5,7 @@
 package com.wynntils.core.chat.tabs;
 
 import com.wynntils.handlers.chat.RecipientType;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
 import com.wynntils.mc.event.ClientsideMessageEvent;
 import java.util.Objects;
 import java.util.Set;
