+++ common/src/main/java/com/wynntils/features/user/ChatTabsFeature.java
@@ -16,7 +16,7 @@ import com.wynntils.core.managers.Models;
 import com.wynntils.gui.widgets.ChatTabAddButton;
 import com.wynntils.gui.widgets.ChatTabButton;
 import com.wynntils.handlers.chat.RecipientType;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
 import com.wynntils.mc.event.ChatScreenKeyTypedEvent;
 import com.wynntils.mc.event.ClientsideMessageEvent;
 import com.wynntils.mc.event.ScreenInitEvent;
