+++ common/src/main/java/com/wynntils/features/user/ChatTimestampFeature.java
@@ -8,7 +8,7 @@ import com.wynntils.core.config.Config;
 import com.wynntils.core.config.ConfigHolder;
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.managers.Model;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
 import com.wynntils.mc.utils.McUtils;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
