+++ common/src/main/java/com/wynntils/features/user/TradeMarketPriceConversionFeature.java
@@ -5,7 +5,7 @@
 package com.wynntils.features.user;
 
 import com.wynntils.core.features.UserFeature;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
 import com.wynntils.mc.event.ChatSentEvent;
 import com.wynntils.mc.event.ScreenOpenedEvent;
 import com.wynntils.mc.utils.McUtils;
