+++ common/src/main/java/com/wynntils/features/user/TranslationFeature.java
@@ -12,8 +12,8 @@ import com.wynntils.core.managers.Model;
 import com.wynntils.core.managers.Models;
 import com.wynntils.core.services.TranslationModel;
 import com.wynntils.handlers.chat.RecipientType;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
-import com.wynntils.handlers.chat.events.NpcDialogEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.NpcDialogEvent;
 import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.mc.utils.McUtils;
 import java.util.List;
