+++ common/src/main/java/com/wynntils/features/user/overlays/NpcDialogueOverlayFeature.java
@@ -26,7 +26,7 @@ import com.wynntils.gui.render.RenderUtils;
 import com.wynntils.gui.render.TextRenderSetting;
 import com.wynntils.gui.render.TextRenderTask;
 import com.wynntils.gui.render.VerticalAlignment;
-import com.wynntils.handlers.chat.events.NpcDialogEvent;
+import com.wynntils.handlers.chat.event.NpcDialogEvent;
 import com.wynntils.mc.event.RenderEvent;
 import com.wynntils.mc.objects.CommonColors;
 import com.wynntils.mc.utils.ComponentUtils;
