+++ common/src/main/java/com/wynntils/features/user/overlays/ObjectivesOverlayFeature.java
@@ -27,6 +27,7 @@ import com.wynntils.handlers.scoreboard.events.ScoreboardSegmentAdditionEvent;
 import com.wynntils.mc.event.RenderEvent;
 import com.wynntils.mc.objects.CommonColors;
 import com.wynntils.mc.objects.CustomColor;
+import com.wynntils.handlers.scoreboard.event.ScoreboardSegmentAdditionEvent;
 import com.wynntils.wynn.model.scoreboard.ScoreboardModel;
 import com.wynntils.wynn.model.scoreboard.objectives.ObjectiveListener;
 import com.wynntils.wynn.model.scoreboard.objectives.WynnObjective;
