+++ common/src/main/java/com/wynntils/handlers/bossbar/event/BossBarAddedEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.handlers.bossbar.events;
+package com.wynntils.handlers.bossbar.event;
 
 import com.wynntils.handlers.bossbar.TrackedBar;
 import net.minecraftforge.eventbus.api.Cancelable;
