+++ common/src/main/java/com/wynntils/handlers/chat/ChatHandler.java
@@ -7,8 +7,8 @@ package com.wynntils.handlers.chat;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.features.Feature;
 import com.wynntils.core.managers.Handler;
-import com.wynntils.handlers.chat.events.ChatMessageReceivedEvent;
-import com.wynntils.handlers.chat.events.NpcDialogEvent;
+import com.wynntils.handlers.chat.event.ChatMessageReceivedEvent;
+import com.wynntils.handlers.chat.event.NpcDialogEvent;
 import com.wynntils.mc.event.ChatPacketReceivedEvent;
 import com.wynntils.mc.objects.ChatType;
 import com.wynntils.mc.utils.ComponentUtils;
