+++ common/src/main/java/com/wynntils/handlers/scoreboard/ScoreboardHandler.java
@@ -6,7 +6,7 @@ package com.wynntils.handlers.scoreboard;
 
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.Handler;
-import com.wynntils.handlers.scoreboard.events.ScoreboardSegmentAdditionEvent;
+import com.wynntils.handlers.scoreboard.event.ScoreboardSegmentAdditionEvent;
 import com.wynntils.mc.event.ScoreboardSetScoreEvent;
 import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.mc.utils.McUtils;
