+++ common/src/main/java/com/wynntils/handlers/scoreboard/event/ScoreboardSegmentAdditionEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.handlers.scoreboard.events;
+package com.wynntils.handlers.scoreboard.event;
 
 import com.wynntils.core.events.EventThread;
 import com.wynntils.handlers.scoreboard.Segment;
